requirements.txt
Jinja2==3.1.6
aiosqlite==0.21.0
bcrypt==4.3.0
databases[aiosqlite]==0.9.0
fastapi==0.116.1
itsdangerous==2.2.0
orjson==3.11.3
pillow==11.3.0
pydantic==2.11.7
python-dotenv==1.1.1
python-multipart==0.0.20
qrcode==8.2
sqlalchemy==2.0.43
uvicorn[standard]==0.35.0